fix authenticate with JWT
